# Backend/backend/src/main/resources/swagger.yaml
swagger: '2.0'
info:
  description: Beschreibung der Rest-Schnitstelle der Healthcare API
  version: 1.0.0
  title: Healthcare App
  license:
    name: Apache 2.0
    url: 'http://www.apache.org/licenses/LICENSE-2.0.html'
host: 'localhost:8080'
basePath: /v1
...
schemes:
  - http
consumes:
  - application/json
produces:
  - application/json
paths:
...
  /maintainancetasksApp:
    post:
      tags:
        - MaintainanceTask
      summary: Erhalte die ersten 20 offenen tasks
      description: ''
      parameters:
        - name: userName
          in: body
          required: true
          schema:
            type: object
            properties:
              userName:
                type: string
                format: string
              id:
                type: integer
                format: int64
        - name: limit
          in: query
          description: Wie viele Tasks sollen geliefert werden
          required: false
          default: 20
          type: integer
          format: int32
        - name: before
          in: query
          type: string
          format: dateTime
          required: false
          default: now
          description: Vor welchem Datum
        - name: since
          in: query
          type: string
          format: dateTime
          required: false
          default: '1970-01-01'
          description: Seit Wann
        - name: open
          in: query
          type: boolean
          required: false
          default: true
        - name: closed
          in: query
          required: false
          type: boolean
          default: true
        - name: processing
          in: query
          required: false
          type: boolean
          default: true
        - name: own
          in: query
          required: false
          type: boolean
          default: false
          description: Nur eigene tasks anzeigen?
        - name: building
          required: false
          type: integer
          in: query
      responses:
        '200':
          description: erfolgreich
          schema:
            type: array
            items:
              $ref: '#/definitions/MaintainanceTask'
        '400':
          description: error
          schema:
            $ref: '#/definitions/ErrorQuery'
  /maintainancetasks:
    get:
      tags:
        - MaintainanceTask
      summary: Erhalte die ersten 20 offenen tasks
      description: ''
      parameters:
        - name: userName
          in: body
          required: true
          schema:
            type: object
            properties:
              userName:
                type: string
                format: string
              id:
                type: integer
                format: int64
        - name: limit
          in: query
          description: Wie viele Tasks sollen geliefert werden
          required: false
          default: 20
          type: integer
          format: int32
        - name: before
          in: query
          type: string
          format: dateTime
          required: false
          default: now
          description: Vor welchem Datum
        - name: since
          in: query
          type: string
          format: dateTime
          required: false
          default: '1970-01-01'
          description: Seit Wann
        - name: open
          in: query
          type: boolean
          required: false
          default: true
        - name: closed
          in: query
          required: false
          type: boolean
          default: true
        - name: processing
          in: query
          required: false
          type: boolean
          default: true
        - name: own
          in: query
          required: false
          type: boolean
          default: false
          description: Nur eigene tasks anzeigen?
        - name: building
          required: false
          type: integer
          in: query
      responses:
        '200':
          description: erfolgreich
          schema:
            type: array
            items:
              $ref: '#/definitions/MaintainanceTask'
        '400':
          description: error
          schema:
            $ref: '#/definitions/ErrorQuery'
    post:
      tags:
        - MaintainanceTask
      summary: Erstelle einen neuen Task
      description: Task
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/MaintainanceTask'
      parameters:
        - name: body
          in: body
          required: true
          schema:
            $ref: '#/definitions/MaintainanceTask'
    put:
        tags:
          - TransportTask
        summary: Eine reihe MaintainanceTask bearbeiten
        description: Irgendwas
        responses:
          '200':
            description: OK
        parameters:
          - name: body
            in: body
            required: true
            schema:
              type: array
              items:
                $ref: '#/definitions/MaintainanceTask'
  /maintainancetasksApp/size:
    post:
      tags:
        - MaintainanceTask
      summary: Anzahl der Tasks
      description: Anzahl der Tasks
      responses:
        '200':
          description: OK
          schema:
            type: integer
            format: int64
      parameters:
        - name: newTasks
          in: query
          required: true
          type: boolean
          description: Nur die anzahl an neuen Tasks haben seit der letzten abfrage
        - name: userName
          in: body
          required: true
          schema:
            type: object
            properties:
              userName:
                type: string
              id:
                type: integer
                format: int64
  /maintainancetasks/size:
    get:
      tags:
        - MaintainanceTask
      summary: Anzahl der Tasks
      description: Anzahl der Tasks
      responses:
        '200':
          description: OK
          schema:
            type: integer
            format: int64
      parameters:
        - name: newTasks
          in: query
          required: true
          type: boolean
          description: Nur die anzahl an neuen Tasks haben seit der letzten abfrage
        - name: userName
          in: body
          required: true
          schema:
            type: object
            properties:
              userName:
                type: string
              id:
                type: integer
                format: int64
  '/maintainancetasks/{taskid}':
    parameters:
      - name: taskid
        in: path
        type: integer
        format: int64
        required: true
    get:
      tags:
        - MaintainanceTask
      summary: Erhalte einen einzelnen Task
      description: Einzelner Task
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/MaintainanceTask'
        '404':
          description: error
          schema:
            type: object
            example:
              message: Id nicht gefunden
    put:
      tags:
        - MaintainanceTask
      summary: Update einen Task
      description: Eine einzelne Task neue Werte zuweisen
      responses:
        '200':
          description: OK
      parameters:
        - name: body
          in: body
          required: true
          schema:
            $ref: '#/definitions/MaintainanceTask'
    delete:
        tags:
          - MaintainanceTask
        summary: löschen
        description: löschen
        responses:
          '200':
            description: OK
          '402':
            description: NOT_FOUND
...
  /beacons:
    get:
      tags:
        - beacon
      summary: Erhalte alle Beacons
      description: Alle Beacons
      parameters:
        - name: noBeaconObject
          in: query
          required: false
          type: boolean
          default: false
          description: Liefert alle beacons die an keinem BeaconObject befestigt sind
      responses:
        '200':
          description: OK
          schema:
            type: array
            items:
              $ref: '#/definitions/Beacon'
    post:
      tags:
        - beacon
      summary: Erstelle neuen Beacon als rückgabe gibt es den fertigen beacon mit uuid
      description: Neuen Beacon erstelen
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/Beacon'
      parameters:
        - name: body
          in: body
          required: true
          schema:
            type: object
            properties:
              major:
                type: integer
                format: int32
              minor:
                type: integer
                format: int32
  '/beacons/{beaconId}':
    parameters:
      - name: beaconId
        in: path
        required: true
        type: string
    get:
      tags:
        - beacon
      summary: Einen bestimmten Beacon erhalten
      description: Einen Bestimmten beacon erhalten
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/Beacon'
    put:
      tags:
        - beacon
      summary: Einen Beacon bearbeiten
      description: Einene Beacon bearbeiten
      responses:
        '200':
          description: OK
      parameters:
        - name: body
          in: body
          required: true
          schema:
            $ref: '#/definitions/Beacon'
    delete:
        tags:
          - Beacon
        summary: löschen
        description: löschen
        responses:
          '200':
            description: OK
          '402':
            description: NOT_FOUND
...
definitions:
  ErrorQuery:
    type: object
    properties:
      message:
        type: string
        format: string
      value:
        type: string
        format: string
        description: Enthält bei falscher Query Parameter den Namen der Query
    example:
      message: Error in der Query
      value: state
  MaintainanceTask:
    allOf:
      - $ref: '#/definitions/Task'
      - properties:
          picture:
            $ref: '#/definitions/PictureOfObject'
          subTasks:
            type: array
            items:
              $ref: '#/definitions/AbstractSubTask'
          repeatTaskInDay:
            type: integer
            format: int32
...
  Location:
    type: object
    properties:
      id:
        type: integer
        format: int64
      building:
        type: integer
        format: int32
      floor:
        type: integer
        format: int32
      room:
        type: integer
        format: int32
    required:
      - building
      - room
      - floor
  Task:
    type: object
    discriminator: type
    required:
      - level
      - creationTime
      - creator
      - description
      - type
    properties:
      id:
        type: integer
        format: int64
      creator:
        $ref: '#/definitions/User'
      editor:
        $ref: '#/definitions/User'
      description:
        type: string
        format: string
        default: Kein Kommentar
      level:
        type: integer
        format: int32
        default: 1
      creationTime:
        type: string
        format: dateTime
      acceptedTime:
        type: string
        format: dateTime
      closedTime:
        type: string
        format: dateTime
      state:
        type: integer
        enum:
          - 1
          - 2
          - 3
        description: 1 bedeutet offen 2 in bearbeitung 3 geschlossen
      beaconObject:
        $ref: '#/definitions/BeaconObject'
      'type':
        type: string
        enum: ['MaintainanceTask','CleaningTask','TransportTask']
      role:
        $ref: '#/definitions/Role'
      name:
        type: string
        format: string
  User:
    type: object
    required:
      - name
      - password
      - email
      - lastUpdate
      - role
    properties:
      id:
        type: integer
        format: int64
      name:
        type: string
        format: string
      password:
        type: string
        format: password
      email:
        type: string
        format: string
      lastUpdate:
        type: string
        format: dateTime
      role:
        $ref: '#/definitions/Role'
  Role:
    type: object
    required:
      - name
    properties:
      id:
        type: integer
        format: int64
      name:
        type: string
        format: string
  BeaconObject:
    type: object
    properties:
      id:
        type: integer
        format: int64
      name:
        type: string
        format: string
      pictureOfObject:
        $ref: '#/definitions/PictureOfObject'
      beacon:
        $ref: '#/definitions/Beacon'
      location:
        $ref: '#/definitions/Location'
      state:
        type: string
        format: string
        description: Kann benutzt werden um einen status zu verdeutlichen
      beaconObjectType:
        type: string
        enum: ['Standardbett','Seniorenbett','Kinderbett','Brandschutztür']
  Beacon:
    type: object
    properties:
      uuid:
        type: string
        format: string
      major:
        type: integer
        format: int32
      minor:
        type: integer
        format: int32
...
  AbstractSubTask:
    type: object
    discriminator: type
    required:
      - type
    properties:
      id:
        type: integer
        format: int64
      title:
        type: string
        format: string
      'type':
        type: string
        enum: ['SubTaskImage','SubTaskCheckbox','SubTaskSlider']
...
  PictureOfObject:
    properties:
      picture:
        type: string
        format: byte